refactor(logic): split remaining oversized logic tests (Refs #2288)
Follow-up to verification on #2288: split four tests still over the 400-line
repo.logic_test_file_size gate into segment files with unchanged cases.
